+++ src/Tsufeki/Tenkawa/Php/Feature/DocumentSymbols/SymbolDocumentSymbolsProvider.php
@@ -11,12 +11,15 @@ use Tsufeki\Tenkawa\Php\Reflection\Element\Function_;
 use Tsufeki\Tenkawa\Php\Reflection\Element\Method;
 use Tsufeki\Tenkawa\Php\Reflection\Element\Property;
 use Tsufeki\Tenkawa\Server\Document\Document;
+use Tsufeki\Tenkawa\Server\Feature\Capabilities\DocumentSymbolClientCapabilities;
 use Tsufeki\Tenkawa\Server\Feature\Common\Location;
 use Tsufeki\Tenkawa\Server\Feature\Common\Range;
 use Tsufeki\Tenkawa\Server\Feature\Common\SymbolInformation;
 use Tsufeki\Tenkawa\Server\Feature\Common\SymbolKind;
+use Tsufeki\Tenkawa\Server\Feature\DocumentSymbols\DocumentSymbol;
 use Tsufeki\Tenkawa\Server\Feature\DocumentSymbols\DocumentSymbolsProvider;
 use Tsufeki\Tenkawa\Server\Utils\PositionUtils;
+use Tsufeki\Tenkawa\Server\Utils\StringUtils;
 
 class SymbolDocumentSymbolsProvider implements DocumentSymbolsProvider
 {
@@ -38,9 +41,9 @@ class SymbolDocumentSymbolsProvider implements DocumentSymbolsProvider
         GlobalSymbol::CONST_ => 1,
         GlobalSymbol::FUNCTION_ => 2,
         GlobalSymbol::CLASS_ => 3,
-        MemberSymbol::CLASS_CONST => 1,
-        MemberSymbol::PROPERTY => 2,
-        MemberSymbol::METHOD => 3,
+        MemberSymbol::CLASS_CONST => 4,
+        MemberSymbol::PROPERTY => 5,
+        MemberSymbol::METHOD => 6,
     ];
 
     public function __construct(SymbolExtractor $symbolExtractor)
@@ -48,7 +51,7 @@ class SymbolDocumentSymbolsProvider implements DocumentSymbolsProvider
         $this->symbolExtractor = $symbolExtractor;
     }
 
-    public function getSymbols(Document $document): \Generator
+    public function getSymbols(Document $document, DocumentSymbolClientCapabilities $capabilities): \Generator
     {
         if ($document->getLanguage() !== 'php') {
             return [];
@@ -61,6 +64,19 @@ class SymbolDocumentSymbolsProvider implements DocumentSymbolsProvider
 
         /** @var DefinitionSymbol[] $symbols */
         $symbols = yield $this->symbolExtractor->getSymbolsInRange($document, $wholeDocumentRange, DefinitionSymbol::class);
+
+        return $capabilities->hierarchicalDocumentSymbolSupport
+            ? $this->getDefinitionSymbols($symbols)
+            : $this->getSymbolInformation($symbols);
+    }
+
+    /**
+     * @param DefinitionSymbol[] $symbols
+     *
+     * @return SymbolInformation[]
+     */
+    private function getSymbolInformation(array $symbols): array
+    {
         /** @var SymbolInformation[] $symbolInfos */
         $symbolInfos = [];
 
@@ -68,7 +84,7 @@ class SymbolDocumentSymbolsProvider implements DocumentSymbolsProvider
 
         foreach ($symbols as $symbol) {
             $symbolInfo = new SymbolInformation();
-            $symbolInfo->name = $symbol->referencedNames[0];
+            $symbolInfo->name = StringUtils::getShortName($symbol->referencedNames[0]);
 
             switch ($symbol->kind) {
                 case GlobalSymbol::CLASS_:
@@ -79,7 +95,7 @@ class SymbolDocumentSymbolsProvider implements DocumentSymbolsProvider
                 case MemberSymbol::PROPERTY:
                 case MemberSymbol::CLASS_CONST:
                 case MemberSymbol::METHOD:
-                    $symbolInfo->containerName = ltrim($symbol->nameContext->class ?? '', '\\');
+                    $symbolInfo->containerName = StringUtils::getShortName($symbol->nameContext->class ?? '');
                     break;
                 default:
                     continue 2;
@@ -94,12 +110,47 @@ class SymbolDocumentSymbolsProvider implements DocumentSymbolsProvider
             }
 
             $symbolInfo->location = new Location();
-            $symbolInfo->location->uri = $document->getUri();
+            $symbolInfo->location->uri = $symbol->document->getUri();
             $symbolInfo->location->range = $symbol->definitionRange;
             $symbolInfos[] = $symbolInfo;
         }
 
         return $symbolInfos;
+    }
+
+    /**
+     * @param DefinitionSymbol[] $symbols
+     *
+     * @return DocumentSymbol[]
+     */
+    private function getDefinitionSymbols(array $symbols): array
+    {
+        /** @var DocumentSymbol[] $documentSymbols */
+        $documentSymbols = [];
+        /** @var DocumentSymbol|null $class */
+        $class = null;
+
+        $this->sort($symbols);
+
+        foreach ($symbols as $symbol) {
+            $documentSymbol = new DocumentSymbol();
+            $documentSymbol->name = StringUtils::getShortName($symbol->referencedNames[0]);
+            $documentSymbol->kind = self::KINDS[$symbol->kind];
+            $documentSymbol->range = $symbol->definitionRange;
+            $documentSymbol->selectionRange = $symbol->range;
+
+            if (in_array($symbol->kind, MemberSymbol::KINDS, true) && $class !== null) {
+                $class->children[] = $documentSymbol;
+            } else {
+                $documentSymbols[] = $documentSymbol;
+            }
+
+            if ($symbol->kind === GlobalSymbol::CLASS_) {
+                $class = $documentSymbol;
+            }
+        }
+
+        return $documentSymbols;
     }
 
     private function isConstructor(DefinitionSymbol $symbol): bool