+++ src/Tsufeki/Tenkawa/Server/Feature/Capabilities/ClientCapabilities.php
@@ -10,4 +10,18 @@ class ClientCapabilities
      * @var WorkspaceClientCapabilities|null
      */
     public $workspace;
+
+    /**
+     * Text document specific client capabilities.
+     *
+     * @var TextDocumentClientCapabilities|null
+     */
+    public $textDocument;
+
+    /**
+     * Experimental client capabilities.
+     *
+     * @var mixed|null
+     */
+    public $experimental;
 }