+++ src/Tsufeki/Tenkawa/Server/Feature/DocumentSymbols/DocumentSymbolsProvider.php
@@ -3,12 +3,13 @@
 namespace Tsufeki\Tenkawa\Server\Feature\DocumentSymbols;
 
 use Tsufeki\Tenkawa\Server\Document\Document;
+use Tsufeki\Tenkawa\Server\Feature\Capabilities\DocumentSymbolClientCapabilities;
 use Tsufeki\Tenkawa\Server\Feature\Common\SymbolInformation;
 
 interface DocumentSymbolsProvider
 {
     /**
-     * @resolve SymbolInformation[]
+     * @resolve DocumentSymbol[]|SymbolInformation[]
      */
-    public function getSymbols(Document $document): \Generator;
+    public function getSymbols(Document $document, DocumentSymbolClientCapabilities $capabilities): \Generator;
 }